tests/wpt/web-platform-tests/css/css-cascade/parsing/all-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Cascading and Inheritance Level 3: parsing all with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-cascade-3/#propdef-all">
<meta name="assert" content="all supports the full grammar 'initial | inherit | unset'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("all", "initial");
test_valid_value("all", "inherit");
test_valid_value("all", "unset");
</script>
</body>
</html>
